Reorganize code into new runtime package
