+++ pkg/runtime/doc.go
@@ -1,0 +1,2 @@
+// Package runtime implements functions to interact with the execution runtime
+package runtime
