+++ pkg/runtime/executor_test.go
@@ -1,4 +1,4 @@
-package process
+package runtime
 
 import (
 	"testing"
