+++ pkg/runtime/fake.go
@@ -1,4 +1,4 @@
-package process
+package runtime
 
 import (
 	"strings"
