+++ pkg/runtime/lock_test.go
@@ -1,4 +1,4 @@
-package process
+package runtime
 
 import (
 	"fmt"
